+++ CHANGELOG.md
@@ -11,8 +11,27 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Compatible changes
 
-- move gemfiles to project root
+- Move gemfiles to project root
+- Added support to read the `include` option from the `travis.yml` file.
 
+  Example:
+
+  ```
+  rvm:
+    - 2.1.8
+    - 2.3.1
+
+  gemfile:
+    - gemfiles/Gemfile1
+    - gemfiles/Gemfile2
+
+  matrix:
+    include:
+      - rvm: 2.1.8
+        gemfile: gemfiles/Gemfile1
+      - rvm: 2.3.1
+        gemfile: gemfiles/Gemfile2
+  ```
 
 ## 0.3.4 - 2018-08-29